+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## [3.1.0] - 2019-06-25
+## [3.X.X] - 2019-X-X
 
 ### Added
 
-- Added init by ID to expandable component
+### Changed
+
+### Fixed
+
+### Removed