+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## [3.1.1] - 2019-06-25
+## [3.X.X] - 2019-X-X
+
+### Added
 
 ### Changed
-- Changed action-link arrow to be material-icons instead of plain text.
+
+### Fixed
+
+### Removed