+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- Removed box-shadow from steps component.
+
 ### Fixed
 
 ### Removed