+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Added .page-footer-rights class to use with .page-footer
+
 ### Changed
 
 ### Fixed