+++ CHANGELOG.md
@@ -12,4 +12,6 @@
 
 ### Fixed
 
+- Fixed issue with px.sheet.open and px.sheet.close
+
 ### Removed