+++ CHANGELOG.md
@@ -4,14 +4,18 @@
 
 ### Added
 
-- Added .page-footer-rights class to use with .page-footer
+- Added .page-footer-rights class to use with .page-footer.
 
 ### Changed
 
-- Removed box-shadow from steps component.
+- There is no longer a width constraint on the card component. Use grid to set sizing.
+- Changed front page and general documentation styling.
 
 ### Fixed
 
-- Fixed issue with px.sheet.open and px.sheet.close
+- Fixed issue with px.sheet.open and px.sheet.close.
 
 ### Removed
+
+- Removed box-shadow from steps component.
+- Removed .cards wrapper.