+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 ### Added
 
 - Added .page-footer-rights class to use with .page-footer.
+- Added better feedback when initializing expandables/accordions
 
 ### Changed
 
 - There is no longer a width constraint on the card component. Use grid to set sizing.
 - Changed front page and general documentation styling.
+- Replaced documentation occurences of PayEx with Swedbank Pay.
 
 ### Fixed